Added a check to the emailservice so that if the template in jasperserver is not found it will log an error.  Also loaded the custom fields into the picklistitem when getting a picklistitem by defaultvalue.
git-svn-id: svn://svn.orangeleap.com/tangerine/tangerine/trunk@7186 0f581344-5646-3a48-8ded-616bfea576c6

+++ src/com/orangeleap/tangerine/service/communication/EmailService.java
@@ -165,27 +165,31 @@ public class EmailService implements ApplicationContextAware {
 
         File tempFile = runReport();
 
-        FileSystemResource file = new FileSystemResource(tempFile);
-        try {
-            helper.addAttachment(getTemplateName() + ".pdf", file);
-            helper.setText("Thank you for your recent donation!");
-            helper.setSubject(subject);
-            helper.setFrom(site.getSmtpFromAddress());
+        if (tempFile != null){
+        	FileSystemResource file = new FileSystemResource(tempFile);
+            try {
+                helper.addAttachment(getTemplateName() + ".pdf", file);
+                helper.setText("Thank you for your recent donation!");
+                helper.setSubject(subject);
+                helper.setFrom(site.getSmtpFromAddress());
 
-            String emailAddresses[] = addresses.split(",");
-            for (int i = 0; i < emailAddresses.length; i++)
-                helper.addTo(emailAddresses[i]);
+                String emailAddresses[] = addresses.split(",");
+                for (int i = 0; i < emailAddresses.length; i++)
+                    helper.addTo(emailAddresses[i]);
 
-            //
-            // finally we mail the message
-            sender.send(message);
+                //
+                // finally we mail the message
+                sender.send(message);
 
 
-            tempFile.delete();
-        } catch (MessagingException e1) {
-            logger.error(e1.getMessage());
-            return;
-        }
+                tempFile.delete();
+            } catch (MessagingException e1) {
+                logger.error(e1.getMessage());
+                return;
+            }
+
+        }else
+        	logger.error("Failed to generate report.");
 
 
     }
